Add granular client tests and update signature logic
Split ClientTest.php into multiple focused test classes for documents, checks, bank statements, business cards, line items, tags, and general client functionality. Updated GenerateSignature trait to handle CURLFile and object payloads more robustly. Upgraded phpunit/phpunit to version 12.3.0 in composer.json.
