+++ composer.json
@@ -25,7 +25,7 @@
   },
 
   "require-dev": {
-    "phpunit/phpunit": "9.5.18"
+    "phpunit/phpunit": "12.3.0"
   },
 
   "autoload": {
