+++ tests/ClientAnyDocsTest.php
@@ -1,0 +1,164 @@
+<?php
+
+use veryfi\client;
+
+require_once __DIR__ . '/ClientTestCase.php';
+
+class ClientAnyDocsTest extends ClientTestCase
+{
+    public function test_process_any_document_base64(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/processAnyDocument.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->atLeastOnce())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+        }
+
+        $file = $this->any_doc_path;
+        $json_response = json_decode($veryfi_client->process_any_document_base64($file, 'us_driver_license'), true);
+        $this->assertNotEmpty($json_response['id']);
+    }
+
+    public function test_process_any_document(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/processAnyDocument.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->atLeastOnce())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+        }
+
+        $file = $this->any_doc_path;
+        $json_response = json_decode($veryfi_client->process_any_document($file, 'us_driver_license'), true);
+        $this->assertNotEmpty($json_response['id']);
+    }
+
+    public function test_process_any_document_url(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/processAnyDocument.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->once())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+        }
+
+        $url = 'https://cdn-dev.veryfi.com/testing/veryfi-python/driver_license.png';
+        $json_response = json_decode($veryfi_client->process_any_document_url($url, 'us_driver_license'), true);
+        $this->assertNotEmpty($json_response['id']);
+    }
+
+    public function test_get_any_documents(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/getAnyDocuments.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->atLeastOnce())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+        }
+
+        $json_response = json_decode($veryfi_client->get_any_documents(), true);
+        $json_len = sizeof($json_response);
+        $this->assertTrue($json_len > 1);
+    }
+
+    public function test_get_any_document(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/getAnyDocuments.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->atLeastOnce())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+        }
+
+
+        if ($this->mock_responses) {
+            $document_id = 12345;
+        } else {
+            $json_response = json_decode($veryfi_client->get_any_documents(), true);
+            $document_id = $json_response['results'][0]['id'];
+        }
+
+        $json_response = json_decode($veryfi_client->get_any_document($document_id), true);
+        $this->assertNotEmpty($json_response);
+    }
+
+    public function test_delete_any_document(): void
+    {
+        if ($this->mock_responses) {
+            $veryfi_client = $this->getMockBuilder(Client::class)
+                ->onlyMethods(['exec_curl'])
+                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
+                ->getMock();
+
+            $file_path = __DIR__ . '/resources/deleteDocument.json';
+            $file = fopen($file_path, 'r');
+            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
+            $veryfi_client->expects($this->once())
+                ->method('exec_curl')
+                ->willReturn($file_data);
+
+            $document_id = 12345;
+
+        } else {
+            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
+
+            $file = $this->any_doc_path;
+            $json_response = json_decode($veryfi_client->process_any_document($file, 'us_driver_license'), true);
+            $document_id = $json_response['id'];
+        }
+
+        $delete_json_response = json_decode($veryfi_client->delete_any_document($document_id));
+        $this->assertEquals(json_decode('{"status": "ok", "message": "Document has been deleted"}'), $delete_json_response);
+    }
+}
