--- tests/ClientTest.php
@@ -1,751 +1,0 @@
-<?php
-
-use PHPUnit\Framework\TestCase;
-use veryfi\client;
-use veryfi\documents\lineitems\LineItem;
-use veryfi\documents\lineitems\LineItemUpdate;
-
-
-final class ClientTest extends TestCase
-{
-    private string $client_id = 'your_client_id';
-    private string $client_secret = 'your_client_secret';
-    private string $username = 'your_username';
-    private string $api_key = 'your_api_key';
-    private string $receipt_path = __DIR__ . '/resources/receipt.jpg';
-    private string $w2_path = __DIR__ . '/resources/w2.png';
-    private string $any_doc_path = __DIR__ . '/resources/driver_license.png';
-    private string $bank_statement_path = __DIR__ . '/resources/bankstatement.pdf';
-    private bool $mock_responses = true;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->client_id = getenv('VERYFI_CLIENT_ID');
-        $this->client_secret = getenv('VERYFI_CLIENT_SECRET');
-        $this->username = getenv('VERYFI_USERNAME');
-        $this->api_key = getenv('VERYFI_API_KEY');
-    }
-
-    public function test_get_documents(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getDocuments.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_documents(), true);
-        $json_len = sizeof($json_response);
-        $this->assertEquals(2, $json_len);
-    }
-
-    public function test_get_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-            $document_id = 125661908;
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-            $documents = json_decode($veryfi_client->get_documents(), true);
-            $document_id = $documents['documents'][0]['id'];
-        }
-        $json_response = json_decode($veryfi_client->get_document($document_id), true);
-        $this->assertEquals($document_id, $json_response['id']);
-    }
-
-    public function test_process_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ .'/resources/processDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $categories = array('Job Supplies');
-        $file = $this->receipt_path;
-        $json_response = json_decode($veryfi_client->process_document_base64($file, $categories, true), true);
-        $this->assertEquals(strtolower('walgreens'), strtolower($json_response['vendor']['name']));
-    }
-
-    private function generate_random_string(): string
-    {
-        $characters = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-        $charactersLength = strlen($characters);
-        $randomString = '';
-        for ($i = 0; $i < 10; $i++) {
-            $randomString .= $characters[rand(0, $charactersLength - 1)];
-        }
-        return $randomString;
-    }
-
-    public function test_update_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/updateDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-            $notes = 'Note updated';
-            $parameters = array('notes' => $notes);
-            $document_id = 125661908;
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-            $documents = json_decode($veryfi_client->get_documents(), true);
-            $document_id = $documents['documents'][0]['id'];
-            $notes = $this->generate_random_string();
-            $parameters = array('notes' => $notes);
-        }
-        $json_response = json_decode($veryfi_client->update_document($document_id, $parameters), true);
-        $this->assertEquals($notes, $json_response['notes']);
-    }
-
-    public function test_delete_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl', 'process_document'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/deleteDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-            $file_path = __DIR__ .'/resources/processDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('process_document')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $categories = array('Job Supplies');
-        $file = $this->receipt_path;
-        $json_response = json_decode($veryfi_client->process_document($file, $categories), true);
-        $id = $json_response['id'];
-        $delete_json_response = json_decode($veryfi_client->delete_document($id));
-        $this->assertEquals(json_decode('{"status": "ok", "message": "Document has been deleted"}'), $delete_json_response);
-    }
-
-    public function test_process_document_url(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/processDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $url = 'https://raw.githubusercontent.com/veryfi/veryfi-python/master/tests/assets/receipt_public.jpg';
-        $json_response = json_decode($veryfi_client->process_document_url($url, null, null, true, 1), true);
-        $this->assertEquals(strtolower('walgreens'), strtolower($json_response['vendor']['name']));
-    }
-
-    public function test_get_line_items(): void
-    {
-        $document_id = 44691518;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getLineItems.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_line_items($document_id), true);
-        $this->assertNotEmpty($json_response['line_items']);
-    }
-
-    public function test_get_line_item(): void
-    {
-        $document_id = 44691518;
-        $line_item_id = 101170751;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getLineItem.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_line_item($document_id, $line_item_id), true);
-        $this->assertNotEmpty($json_response);
-        $this->assertEquals($line_item_id, $json_response['id']);
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function test_update_line_item(): void
-    {
-        $document_id = 44691518;
-        $line_item_id = 101170751;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/updateLineItem.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $fields_to_update = array('description' => 'TEST');
-        $fields_to_update = new LineItemUpdate($fields_to_update);
-        $json_response = json_decode($veryfi_client->update_line_item($document_id, $line_item_id, $fields_to_update), true);
-        $this->assertEquals('TEST', $json_response['description']);
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function test_add_line_item(): void
-    {
-        $document_id = 44691518;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/addLineItem.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $line_item = array('order' => 20, 'description' => 'TEST', 'total' => 20.1, 'sku' => 'aqw');
-        $line_item = new LineItem($line_item);
-        $json_response = json_decode($veryfi_client->add_line_item($document_id, $line_item), true);
-        $this->assertEquals($line_item->order, $json_response['order']);
-        $this->assertEquals($line_item->description, $json_response['description']);
-        $this->assertEquals($line_item->total, $json_response['total']);
-        $this->assertEquals($line_item->sku, $json_response['sku']);
-        if (!$this->mock_responses) {
-            $id = $json_response['id'];
-            $json_response = json_decode($veryfi_client->delete_line_item($document_id, $id), true);
-            $this->assertEquals('ok', $json_response['status']);
-        }
-    }
-
-    public function test_delete_line_item(): void
-    {
-        $document_id = 44691518;
-        $line_item_id = 189951682;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/deleteLineItem.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-            $json_response = json_decode($veryfi_client->delete_line_item($document_id, $line_item_id), true);
-            $this->assertEquals('ok', $json_response['status']);
-        } else {
-            $this->assertTrue(true); // Tested before
-        }
-    }
-
-    public function test_delete_line_items(): void
-    {
-        $document_id = 51208553;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/deleteLineItems.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->delete_line_items($document_id), true);
-        $this->assertEquals('ok', $json_response['status']);
-    }
-
-    public function test_validate_signature(): void
-    {
-        $client_signature = "m89UF6aTlce2YcVbGw5LTZuA+bc5MPVS9AOicjkS7qM=";
-        $client_secret = "fAKEB2oJMLbHwBN5jEd6h3f3Lj1o9gK5kcz2xAf8Kyi2X1PNaJ6F612344YcOsSllGkFAkeUiZV5ZTNoPkk6bXyctGGAdfcratu4Dl2CA2XtU6En5icHxjVRUNoSFGP";
-        $payload = array("event" => "document.created", "data" => array("id" => 63184393, "created" => "2022-03-28 21:12:14"));
-        $this->assertTrue(Client::verify_signature($payload["data"], $client_secret, $client_signature));
-    }
-
-    public function test_bad_credentials(): void
-    {
-        $veryfi_client = new Client('', '', '', '');
-        $json_response = json_decode($veryfi_client->get_documents(), true);
-        $this->assertEquals('fail', $json_response['status']);
-    }
-
-    public function test_add_tag(): void
-    {
-        $document_id = 140512586;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/addTag.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->add_tag($document_id, 'test_tag'), true);
-        $this->assertEqualsIgnoringCase('test_tag', $json_response['name']);
-        if (!$this->mock_responses) {
-            $json_response = json_decode($veryfi_client->delete_tags($document_id), true);
-            $this->assertEmpty($json_response);
-        }
-    }
-
-    public function test_delete_tags(): void
-    {
-        $document_id = 140512586;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/deleteTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->delete_tags($document_id), true);
-        $this->assertEmpty($json_response);
-    }
-
-    public function test_get_tags(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_tags(), true);
-        $this->assertNotEmpty($json_response);
-    }
-
-    public function test_get_document_tags(): void
-    {
-        $document_id = 140804210;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_document_tags($document_id), true);
-        $this->assertNotEmpty($json_response);
-    }
-
-    public function test_delete_tag(): void
-    {
-        $document_id = 140512586;
-        $tag_id = 6673689;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/deleteTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->delete_tag($document_id, $tag_id), true);
-        $this->assertEmpty($json_response);
-    }
-
-    public function test_replace_document_tags(): void
-    {
-        $document_id = 140804210;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $tags = array('TAG_1', 'TAG_2', 'TAG_3');
-        $json_response = json_decode($veryfi_client->replace_tags($document_id, $tags), true);
-        $this->assertNotEmpty($json_response);
-    }
-
-    public function test_add_document_tags(): void
-    {
-        $document_id = 140804210;
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getTags.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $tags = array('TAG_1', 'TAG_2', 'TAG_3');
-        $json_response = json_decode($veryfi_client->add_tags($document_id, $tags), true);
-        $this->assertNotEmpty($json_response);
-    }
-
-    public function test_process_w2_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ .'/resources/processW2Document.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $file = $this->w2_path;
-        $json_response = json_decode($veryfi_client->process_w2_base64($file, true), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    public function test_process_any_document(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ .'/resources/processAnyDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->atLeastOnce())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $file = $this->any_doc_path;
-        $json_response = json_decode($veryfi_client->process_any_document_base64($file, 'us_driver_license'), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    public function test_process_bank_statement(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ .'/resources/processBankStatement.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->atLeastOnce())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $file = $this->bank_statement_path;
-        $json_response = json_decode($veryfi_client->process_bank_statement_base64($file), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    public function test_process_any_document_url(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/processAnyDocument.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $url = 'https://cdn-dev.veryfi.com/testing/veryfi-python/driver_license.png';
-        $json_response = json_decode($veryfi_client->process_any_document_url($url, 'us_driver_license'), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    public function test_process_bank_statement_url(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/processBankStatement.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $url = 'https://cdn-dev.veryfi.com/testing/veryfi-python/bankstatement.pdf';
-        $json_response = json_decode($veryfi_client->process_bank_statement_url($url), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    public function test_process_w2_document_from_url(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-            $file_path = __DIR__ . '/resources/processW2Document.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->once())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $file_name = 'w2_form.pdf';
-        $url = 'https://cdn.veryfi.com/wp-content/uploads/image.png';
-        $json_response = json_decode($veryfi_client->process_w2_from_url($file_name, $url, null, true), true);
-        $this->assertNotEmpty( $json_response['id']);
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function test_get_w2_documents(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-            $file_path = __DIR__ . '/resources/getW2Documents.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->atLeastOnce())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-        $json_response = json_decode($veryfi_client->get_w2s(), true);
-        $json_len = sizeof($json_response);
-        $this->assertTrue($json_len > 1);
-        $document_id = $json_response['results'][0]['id'];
-        $json_response = json_decode($veryfi_client->get_w2($document_id), true);
-        $this->assertNotEmpty( $json_response);
-    }
-
-    public function test_get_bank_statements(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getBankStatements.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->atLeastOnce())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $json_response = json_decode($veryfi_client->get_bank_statements(), true);
-        $json_len = sizeof($json_response);
-        $this->assertTrue($json_len > 1);
-        $document_id = $json_response['results'][0]['id'];
-        $json_response = json_decode($veryfi_client->get_bank_statement($document_id), true);
-        $this->assertNotEmpty( $json_response);
-    }
-
-    public function test_get_any_documents(): void
-    {
-        if ($this->mock_responses) {
-            $veryfi_client = $this->getMockBuilder(Client::class)
-                ->onlyMethods(['exec_curl'])
-                ->setConstructorArgs([$this->client_id, $this->client_secret, $this->username, $this->api_key])
-                ->getMock();
-
-            $file_path = __DIR__ . '/resources/getAnyDocuments.json';
-            $file = fopen($file_path, 'r');
-            $file_data = mb_convert_encoding(fread($file, filesize($file_path)), 'UTF-8');
-            $veryfi_client->expects($this->atLeastOnce())
-                ->method('exec_curl')
-                ->willReturn($file_data);
-
-        } else {
-            $veryfi_client = new Client($this->client_id, $this->client_secret, $this->username, $this->api_key);
-        }
-
-        $json_response = json_decode($veryfi_client->get_any_documents(), true);
-        $json_len = sizeof($json_response);
-        $this->assertTrue($json_len > 1);
-        $document_id = $json_response['results'][0]['id'];
-        $json_response = json_decode($veryfi_client->get_any_document($document_id), true);
-        $this->assertNotEmpty( $json_response);
-    }
-}
