+++ tests/resources/getW8BenEs.json
@@ -1,0 +1,12 @@
+{
+    "documents": [
+        {
+            "id": 1122334455,
+            "document_type": "w8bene",
+            "created_date": "2023-01-01 12:00:00"
+        }
+    ],
+    "meta": {
+        "total_results": 1
+    }
+}
