+++ tests/resources/getW9s.json
@@ -1,0 +1,12 @@
+{
+    "documents": [
+        {
+            "id": 5544332211,
+            "document_type": "w9",
+            "created_date": "2023-01-01 12:00:00"
+        }
+    ],
+    "meta": {
+        "total_results": 1
+    }
+}
