+++ tests/resources/processCheck.json
@@ -1,0 +1,6 @@
+{
+    "id": 987654321,
+    "document_type": "check",
+    "created_date": "2023-01-01 12:00:00",
+    "total": 100.00
+}
