+++ tests/resources/processW8BenE.json
@@ -1,0 +1,5 @@
+{
+    "id": 1122334455,
+    "document_type": "w8bene",
+    "created_date": "2023-01-01 12:00:00"
+}
