+++ tests/resources/processW9.json
@@ -1,0 +1,5 @@
+{
+    "id": 5544332211,
+    "document_type": "w9",
+    "created_date": "2023-01-01 12:00:00"
+}
